requirements.txt
# ── Python >= 3.11 required ────────────────────────────────────────────────
# Preferred: uv sync
# Fallback:  pip install -r requirements.txt

# Core ML
torch>=2.5.0
torchvision>=0.20.0
timm>=1.0.0
transformers>=4.45.0
pytorch-metric-learning>=2.3.0
accelerate>=0.34.0

# Image processing
Pillow>=10.4.0
opencv-python>=4.10.0
numpy>=1.26.4
scipy>=1.13.1
scikit-learn>=1.5.2
scikit-image>=0.24.0
PyWavelets>=1.7.0
imageio>=2.34.0
albumentations>=1.4.0

# OCR
easyocr>=1.7.1

# Streamlit UI
streamlit>=1.40.0
plotly>=5.22.0
streamlit-image-comparison>=0.0.4

# Utilities
tqdm>=4.66.5
python-dotenv>=1.0.0
huggingface-hub>=0.24.0
requests>=2.31.0

# Notebooks
jupyter>=1.1.0
ipywidgets>=8.1.0
matplotlib>=3.9.0

# Dev (optional — install with: pip install -r requirements.txt -r requirements-dev.txt)
# pytest>=8.0.0
# ruff>=0.6.0
